requirements.txt
backtesting
fastapi
numpy
pandas
PyPortfolioOpt
quantstats
scipy
yfinance
httpx
ipython
pytest-mock
